refactor(tools): split tool registrations into domain modules + add MEMORY_ENABLED
Split the 1665-line tool-definitions.ts monolith into four domain
group modules under mcp-core/tools/:

- vault-crud-tools.ts (9 tools) — read/write/patch/delete/move
- search-tools.ts (11 tools) — search, property, and graph queries
- memory-tools.ts (4 tools) — About Me/ memory layer
- daily-note-tools.ts (1 tool) — daily note resolution
- tool-helpers.ts — shared ToolRegistrationContext type + safeHandler

The orchestrator (tool-definitions.ts) is now ~50 lines: imports
each group, composes TOOL_NAMES, and calls register functions.

Layer the MEMORY_ENABLED env var on top: when false, the orchestrator
skips registerMemoryTools(), server.ts skips bootstrap, mcp-router
strips memory from metadata, prompt-definitions hides memory-review
and adjusts vault-orientation + daily-review. Default true preserves
existing behavior.
