+++ src/vault-mcp/mcp-core/mcp-router.ts
@@ -82,12 +82,18 @@ export const createMcpRouter = ({
             name: "vault-cortex",
             title: "Vault Cortex",
             version: "1.0.0",
-            description: `Read, write, and search an Obsidian vault. Provides full-text search, tag queries, and a structured memory layer (${config.memoryDir}/) for personalization across conversations.`,
+            description: config.memoryEnabled
+              ? `Read, write, and search an Obsidian vault. Provides full-text search, tag queries, and a structured memory layer (${config.memoryDir}/) for personalization across conversations.`
+              : `Read, write, and search an Obsidian vault. Provides full-text search, tag queries, and property-based filtering.`,
             icons: SERVER_ICONS,
             websiteUrl: SERVER_WEBSITE_URL,
           },
           {
-            instructions: `Read, write, and search an Obsidian vault. Use vault_search and vault_read_note to find and read notes. Use vault_get_memory to retrieve user preferences and context from ${config.memoryDir}/ files. Use vault_write_note and vault_update_memory for writes.
+            instructions: config.memoryEnabled
+              ? `Read, write, and search an Obsidian vault. Use vault_search and vault_read_note to find and read notes. Use vault_get_memory to retrieve user preferences and context from ${config.memoryDir}/ files. Use vault_write_note and vault_update_memory for writes.
+
+Vault content is Obsidian Flavored Markdown. Write tools pass content through without escaping — be intentional about Obsidian syntax (#, [[, %%, etc.) in inputs.`
+              : `Read, write, and search an Obsidian vault. Use vault_search and vault_read_note to find and read notes. Use vault_write_note for writes.
 
 Vault content is Obsidian Flavored Markdown. Write tools pass content through without escaping — be intentional about Obsidian syntax (#, [[, %%, etc.) in inputs.`,
           },
