+++ src/vault-mcp/server.ts
@@ -73,8 +73,12 @@ const startServer = async (): Promise<void> => {
   const count = await search.rebuildFromVault(vaultPath)
   logger.info("initial index built", { count })
 
-  const memoryStore = createMemoryStore({ memoryDir: config.memoryDir })
-  await memoryStore.bootstrapMemoryDir({ vaultPath }, logger)
+  if (config.memoryEnabled) {
+    const memoryStore = createMemoryStore({ memoryDir: config.memoryDir })
+    await memoryStore.bootstrapMemoryDir({ vaultPath }, logger)
+  } else {
+    logger.info("memory layer disabled")
+  }
 
   if (config.windowsBindMount) {
     logger.info("windows bind-mount mode enabled", {
